requirements.txt
Django~=5.0.1
Jinja2~=3.1.3
asgiref~=3.7.2
sqlparse~=0.4.4
django-debug-toolbar~=4.2.0
pytz~=2023.4
djangorestframework~=3.14.0